src/requirements.txt
click==8.1.7
minio==7.2.7
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
python-dotenv==1.0.1
pymongo==4.7.2
pyspark==3.5.0
torch==2.3.0
transformers==4.41.1
sentence-transformers==3.0.0
streamlit==1.35.0
